requirements.txt
ffmpy
flake8
ipython
Pillow
pillow-avif-plugin
python-magic